+++ src/index.js
@@ -1,7 +1,12 @@
+import express from 'express';
 import 'dotenv/config';
-import { app } from './server.js';
+import webContext from './webContext.js';
 
 const PORT = process.env.PORT || 3000;
+
+const app = express();
+
+webContext(app);
 
 app.listen(PORT, () => {
   console.log(`Server started on port ${PORT}`);
